+++ src/main/java/phanes/replay/user/client/KakaoProfileClient.java
@@ -3,7 +3,7 @@ package phanes.replay.user.client;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
-import phanes.replay.user.dto.oauth.KakaoProfileResponse;
+import phanes.replay.user.dto.oauth.kakao.KakaoProfileResponse;
 
 @FeignClient(name = "kakaoProfileClient", url = "https://kapi.kakao.com")
 public interface KakaoProfileClient {
