+++ src/main/java/phanes/replay/user/client/NaverProfileClient.java
@@ -3,7 +3,7 @@ package phanes.replay.user.client;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
-import phanes.replay.user.dto.oauth.NaverProfileResponse;
+import phanes.replay.user.dto.oauth.naver.NaverProfileResponse;
 
 @FeignClient(value = "Naver", url = "https://openapi.naver.com")
 public interface NaverProfileClient {
