--- src/main/java/phanes/replay/user/dto/oauth/KakaoProfileResponse.java
@@ -1,15 +1,0 @@
-package phanes.replay.user.dto.oauth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class KakaoProfileResponse {
-    private Long id;
-    private KakaoProfile kakao_account;
-}
