+++ src/main/java/phanes/replay/user/dto/oauth/google/GoogleTokenResponse.java
@@ -1,4 +1,4 @@
-package phanes.replay.user.dto.oauth;
+package phanes.replay.user.dto.oauth.google;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
